# application/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于区块链技术的房地产交易系统api文档",
        "title": "基于区块链技术的房地产交易系统api文档",
        "contact": {
            "name": "夜央 Oh oh oh oh oh oh",
            "email": "zoujh99@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/v1/createRealEstate": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "新建房地产(管理员)",
                "parameters": [
                    {
                        "description": "realEstate",
                        "name": "realEstate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.RealEstateRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/hello": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "测试输出Hello",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/queryAccountList": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取账户信息",
                "parameters": [
                    {
                        "description": "account",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.AccountRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "v1.AccountIdBody": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string"
                }
            }
        },
        "v1.AccountRequestBody": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.AccountIdBody"
                    }
                }
            }
        },
        "v1.RealEstateRequestBody": {
            "type": "object",
            "properties": {
                "accountId": {
                    "description": "操作人ID",
                    "type": "string"
                },
                "livingSpace": {
                    "description": "生活空间",
                    "type": "number"
                },
                "proprietor": {
                    "description": "所有者(业主)(业主AccountId)",
                    "type": "string"
                },
                "totalArea": {
                    "description": "总面积",
                    "type": "number"
                }
            }
        }
    }
}